+++ tensordict/nn/common.py
@@ -13,6 +13,7 @@ from typing import Any, Callable, Dict, Iterable, List, Optional, Sequence, Tupl
 
 import torch
 from cloudpickle import dumps as cloudpickle_dumps, loads as cloudpickle_loads
+from tensordict._td import is_tensor_collection, TensorDictBase
 from tensordict._tensordict import unravel_key_list
 from tensordict.functional import make_tensordict
 
@@ -25,7 +26,6 @@ from tensordict.nn.functional_modules import (
 )
 
 from tensordict.nn.utils import set_skip_existing
-from tensordict.td import is_tensor_collection, TensorDictBase
 from tensordict.utils import implement_for, NestedKey
 from torch import nn, Tensor
