+++ tensordict/nn/functional_modules.py
@@ -15,9 +15,9 @@ from typing import Any, Callable, Iterable
 
 import torch
 from tensordict._pytree import PYTREE_REGISTERED_TDS
-from tensordict.base import _is_tensor_collection, TensorDictBase
 
-from tensordict.td import TensorDict
+from tensordict._td import TensorDict
+from tensordict.base import _is_tensor_collection, TensorDictBase
 
 from tensordict.utils import implement_for
 from torch import nn
