+++ tensordict/tensorclass.py
@@ -20,10 +20,10 @@ from typing import Any, Callable, Sequence, TypeVar
 import tensordict as tensordict_lib
 
 import torch
+from tensordict._td import is_tensor_collection, NO_DEFAULT, TensorDict, TensorDictBase
 from tensordict._tensordict import _unravel_key_to_tuple
 from tensordict._torch_func import TD_HANDLED_FUNCTIONS
 from tensordict.memmap_deprec import MemmapTensor as _MemmapTensor
-from tensordict.td import is_tensor_collection, NO_DEFAULT, TensorDict, TensorDictBase
 
 from tensordict.utils import (
     _get_repr,
